Fix handoff kick releasing the session and split delegation into its own mixin
