+++ music_assistant/controllers/player_queues/queue_loader.py
@@ -45,13 +45,13 @@ from music_assistant.controllers.player_queues.autoplay import (
     AUTOPLAY_SERIES_MEDIA_TYPES,
     AutoplayMode,
 )
-from music_assistant.controllers.player_queues.base import _PlayerQueuesBase
 from music_assistant.controllers.player_queues.constants import (
     CONF_DEFAULT_ENQUEUE_OPTION_LIVE_SOURCES,
     MANAGED_POOL_MAX,
     ORDERED_MEDIA_TYPES,
     PROBED_DURATION_MEDIA_TYPES,
 )
+from music_assistant.controllers.player_queues.delegation import PlaybackDelegationMixin
 from music_assistant.controllers.player_queues.helpers import (
     build_queue_item,
     handle_play_action,
@@ -68,14 +68,13 @@ from music_assistant.helpers.throttle_retry import BYPASS_THROTTLER
 from music_assistant.models.music_provider import MusicProvider
 
 if TYPE_CHECKING:
-    from music_assistant_models.media_items import AudioSource
     from music_assistant_models.media_items.metadata import MediaItemImage
     from music_assistant_models.queue_item import QueueItem
 
     from music_assistant.providers.radio_playlist import RadioPlaylistProvider
 
 
-class QueueLoaderMixin(_PlayerQueuesBase):
+class QueueLoaderMixin(PlaybackDelegationMixin):
     """Load items into a queue: apply the enqueue option, resolve single items, refill the pool."""
 
     async def _enqueue_with_option(
@@ -991,142 +990,6 @@ class QueueLoaderMixin(_PlayerQueuesBase):
                 self._ensure_current_index(queue_id)
         finally:
             self._set_transitioning(queue_id, False)
-
-    async def _apply_playback_delegation(
-        self,
-        queue_id: str,
-        media_items: list[MediaItemType],
-        option: QueueOption | None,
-        context: MediaItemType | BrowseFolder | None,
-        start_item: PlayableMediaItemType | str | None,
-    ) -> list[MediaItemType]:
-        """
-        Redirect the play request into an external session when a provider delegates it.
-
-        Returns the media items the normal enqueue should continue with: the delegate
-        AudioSource when the whole batch was redirected into one session, an empty list
-        when the request was handled entirely session-side (enqueue onto the already
-        active session), or the (possibly reduced) original items.
-
-        :param queue_id: The queue the play request targets.
-        :param media_items: The resolved playable items of the play request.
-        :param option: The (settled) enqueue option of the play request.
-        :param context: The single original container the request expanded from, if any.
-        :param start_item: Optional item (or uri) within the context to start at.
-        """
-        if not media_items or option is None:
-            return media_items
-        add_family = option in (QueueOption.ADD, QueueOption.NEXT, QueueOption.REPLACE_NEXT)
-        delegate_cache: dict[str, tuple[AudioSource, MusicProvider] | None] = {}
-        item_delegates = [
-            await self._get_item_delegate(queue_id, item, add_family, delegate_cache)
-            for item in media_items
-        ]
-        if all(delegated is not None for delegated in item_delegates) and (
-            len({delegated[0].uri for delegated in item_delegates if delegated is not None}) == 1
-        ):
-            delegate, provider = cast(
-                "tuple[AudioSource, MusicProvider]",
-                item_delegates[0],
-            )
-            self.logger.info(
-                "Redirecting playback of %d item(s) into external session %s",
-                len(media_items),
-                delegate.name,
-            )
-            await provider.play_on_delegate(
-                delegate,
-                cast("list[PlayableMediaItemType]", media_items),
-                option,
-                queue_id,
-                context=None if isinstance(context, BrowseFolder) else context,
-                start_item=start_item,
-            )
-            if add_family:
-                # the items live in the session's own queue; nothing to load into the
-                # MA queue (the session's AudioSource is already its current item)
-                return []
-            # the session plays the items; the MA queue plays the session
-            return [delegate]
-        # The batch cannot ride one session. Items with a normal playback path (any
-        # available mapping on a provider that streams itself) play through the
-        # regular path; in a mixed batch the delegate-only items are dropped. A batch
-        # where NOTHING plays normally is kept whole, so the delegate-only provider's
-        # own actionable error surfaces at stream time instead of a silent drop.
-        plays_normally = [self._has_normal_playback_path(item) for item in media_items]
-        if all(plays_normally) or not any(plays_normally):
-            return media_items
-        dropped = [
-            item for item, normal in zip(media_items, plays_normally, strict=True) if not normal
-        ]
-        self.logger.warning(
-            "Skipping %d item(s) (%s): mixing session-redirected items with other content "
-            "in one queue is not supported yet - play them separately",
-            len(dropped),
-            ", ".join(item.name for item in dropped[:5]),
-        )
-        return [item for item, normal in zip(media_items, plays_normally, strict=True) if normal]
-
-    def _has_normal_playback_path(self, item: MediaItemType) -> bool:
-        """Return whether any of the item's mappings plays through the regular streaming path."""
-        for mapping in item.provider_mappings:
-            if not mapping.available:
-                continue
-            provider = self.mass.get_provider(mapping.provider_instance)
-            if isinstance(provider, MusicProvider) and not provider.playback_requires_delegate:
-                return True
-        return False
-
-    async def _get_item_delegate(
-        self,
-        queue_id: str,
-        item: MediaItemType,
-        add_family: bool,
-        delegate_cache: dict[str, tuple[AudioSource, MusicProvider] | None],
-    ) -> tuple[AudioSource, MusicProvider] | None:
-        """
-        Return the external session (and owning music provider) that should play the item.
-
-        :param queue_id: The queue the play request targets.
-        :param item: The resolved media item to find a playback delegate for.
-        :param add_family: Whether the request enqueues onto the session (ADD/NEXT family)
-            rather than starting playback, which gates on the enqueueable media types.
-        :param delegate_cache: Per-request cache of each provider's delegate answer.
-        """
-        # stable order: provider_mappings is a set, and every item of the batch must
-        # pick the same delegate when its mappings offer more than one (e.g. a library
-        # item mapped to two accounts of the same service)
-        mappings = sorted(
-            item.provider_mappings,
-            key=lambda mapping: (mapping.provider_domain, mapping.provider_instance),
-        )
-        for mapping in mappings:
-            if not mapping.available:
-                continue
-            if mapping.provider_instance not in delegate_cache:
-                provider = self.mass.get_provider(mapping.provider_instance)
-                if not isinstance(provider, MusicProvider):
-                    delegate_cache[mapping.provider_instance] = None
-                    continue
-                try:
-                    delegate = await provider.get_playback_delegate(queue_id)
-                except Exception as err:
-                    self.logger.warning(
-                        "Error resolving playback delegate from %s: %s", provider.name, err
-                    )
-                    delegate = None
-                delegate_cache[mapping.provider_instance] = (
-                    (delegate, provider) if delegate is not None else None
-                )
-            if (delegated := delegate_cache[mapping.provider_instance]) is None:
-                continue
-            caps = delegated[0].queue_capabilities
-            if caps is None:
-                continue
-            allowed = caps.enqueueable_media_types if add_family else caps.playable_media_types
-            if item.media_type in allowed:
-                return delegated
-        return None
 
     async def _get_similar_tracks(
         self,
